Include memoverride.cpp to fix crashes
GMod seems to use the custom allocator now, and since we're using the
engine's vtables for our own objects, we need to as well (otherwise
there are problems when virtual destructors are called). Ideally I'll
implement a system which copies everything from the engine's vtables and
replaces the destructors with our own, but for now this'll do.